// src/lib/fineract/generated/types.gen.ts
export type ExternalId = {
    empty?: boolean;
    value?: string;
};
// ...
export type Image = {
    id?: number;
    location?: string;
    new?: boolean;
    storageType?: number;
};
// ...
export type Office = {
    children?: Array<Office>;
    externalId?: ExternalId;
    hierarchy?: string;
    id?: number;
    name?: string;
    new?: boolean;
    openingDate?: string;
    openingLocalDate?: string;
    parent?: Office;
};

export type OfficeData = {
    dateFormat?: string;
    externalId?: ExternalId;
    hierarchy?: string;
    id?: number;
    locale?: string;
    name?: string;
    nameDecorated?: string;
    openingDate?: string;
    parentId?: number;
    parentName?: string;
    rowIndex?: number;
};
// ...
export type Staff = {
    active?: boolean;
    displayName?: string;
    emailAddress?: string;
    externalId?: string;
    firstname?: string;
    id?: number;
    image?: Image;
    joiningDate?: string;
    lastname?: string;
    loanOfficer?: boolean;
    mobileNo?: string;
    new?: boolean;
    notActive?: boolean;
    notLoanOfficer?: boolean;
    office?: Office;
    organisationalRoleParentStaff?: Staff;
    organisationalRoleType?: number;
};